app/requirements_migration.txt
asyncpg==0.29.0
gspread==6.1.2
google-auth==2.34.0
python-dotenv==1.0.1
